+++ usbsas-privileges/src/imager.rs
@@ -15,7 +15,7 @@ pub fn drop_priv(fds_read: Vec<RawFd>, fds_write: Vec<RawFd>) -> Result<()> {
         Syscall::ioctl,
         &[
             Comparator::new(0, Cmp::Eq, 1, None),
-            Comparator::new(1, Cmp::Eq, libc::TCGETS as u64, None),
+            Comparator::new(1, Cmp::Eq, libc::TCGETS, None),
         ],
     )?;
     // ioctl(2, TCGETS, ..)
@@ -24,7 +24,7 @@ pub fn drop_priv(fds_read: Vec<RawFd>, fds_write: Vec<RawFd>) -> Result<()> {
         Syscall::ioctl,
         &[
             Comparator::new(0, Cmp::Eq, 2, None),
-            Comparator::new(1, Cmp::Eq, libc::TCGETS as u64, None),
+            Comparator::new(1, Cmp::Eq, libc::TCGETS, None),
         ],
     )?;
     // ioctl(2, TIOCGWINSZ, ..)
@@ -33,7 +33,7 @@ pub fn drop_priv(fds_read: Vec<RawFd>, fds_write: Vec<RawFd>) -> Result<()> {
         Syscall::ioctl,
         &[
             Comparator::new(0, Cmp::Eq, 2, None),
-            Comparator::new(1, Cmp::Eq, libc::TIOCGWINSZ as u64, None),
+            Comparator::new(1, Cmp::Eq, libc::TIOCGWINSZ, None),
         ],
     )?;
